test/reftest/vendor/webkit/fast/dom/Window/window-legacy-event-listener.html

<!DOCTYPE HTML PUBLIC "-//IETF//DTD HTML//EN">
<html>
<head>
<script src="../../../resources/wpt-test-pre.js"></script>
<script src="../../../resources/js-test-pre.js"></script>
</head>
<body>
<script>
//description("Test addEventListener() and removeEventListener() fail silently if arguments are missing.");

shouldThrow('window.addEventListener("foo")');
shouldThrow('window.removeEventListener("bar")');

shouldThrow('window.addEventListener()');
shouldThrow('window.removeEventListener()');
</script>
<script src="../../../resources/js-test-post.js"></script>
<script src="../../../resources/wpt-test-post.js"></script>
</body>
</html>
